[RAPTOR-8303] move more test cases to be run per environment - drop 1/many
